requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
